+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 # [9.8.0] - 2025-01-29
 - Added `trusted_recipient` to SMS, MMS, and MMS
+- Added support for Typing Indicators in WhatsApp
 
 # [9.7.0] - 2025-12-04
 - Added native RCS message types (card, carousel) and suggestions support to eliminate need for custom message type